Create Dictionary
This program is created using Python. What it does is when the user inputs a word, it will return the definition. If the user makes a mistake in typing the word, it will leave suggestions like " Did you mean this instead?". If yes, it will print the definition of the new correct word. Else it will return saying that the word does not exist and will prompt you to double check and type in the correct word. 

This project is part of the Python Mega Course on Udemy.