Website Blocker
Website Blocker app blocks websites that the user specifies on the website list. Once the user specifies the host file and the websites that they would like to block, the app will specify the website on the host file and deny access during the times that are set. Once that time period is over, the user will be allowed to access those websites until the next work period starts. This app uses the concepts of datetime methods, file commands, and conditional statements in Python.