Creating web maps using folium and pandas
This is an app that uses Folium- python library which builds leaflet maps- and pandas- python package that includes data analysis tools- to show population density across the world and volcanoes and it's elevation in the United States. There is a Volcanoes text file which contains all the volcanoes and their elevation in the western part of United States. Pandas package reads the large text file in an efficient and fast way which reduces compile time for the app.  Folium allows the app to add and manipulate features like adding and coloring the markers, zoom into certain coordinates, and coloring the map certain colors based on the density of the population. 
In this app, I have separated Volcanoes and the population as feature group. In the end, I add each feature group onto the map which is more efficient to read but also easy to debug if one feature group does not work.