Books database changes (#1)
* Create frontend.py

This front-end application is created using Tkinter library in python.

* Front end books desktop application

This application is called BookBase which stores information about Books that the user provides. It is split into two parts: front end and back end. This part is the front end. The requirements for this part was to design a GUI which includes  :
1.	Create a GUI window
2.	Add commands like the view, search, add, delete, and update entries. 
3.	Create labels like Title, Author, Year, ISBN and place them using the grid method and allowing text entry into the labels
4.	Creating the Listbox which shows the entries that have been entered and creating a scrollbar on the window.
5.	Creating buttons for commands in step 2
6.	Write a function that returns list object with a tuple that consists of id, title, author, year, ISBN.
7.	Import backend files into front-end

These requirements were achieved by adding 4 entry boxes for the user to enter information which are Title, Author, Year, and ISBN. Also, 8 widgets were added which include: 1.	Listbox: which displays the entries in the table
2.	Scrollbar: used when there are multiple entries in the list box.
3.	View all – shows all the books in the table
4.	Search entry: allows the user to search for the entries based on the information inputted.
5.	Add Entry: adds an entry that the user inputs.
6.	Update Selected: updates an existing entry within the table
7.	Delete Selected: deletes an entry in the table.
8.	Close: closes the application





Credits : 
Python Mega Course by Ardit Sulce.