Back end books database application
This application is called BookBase which stores information about Books that the user provides. It is split into two parts: front end and back end. This part is the back end. The requirements for this part was to design a table using SQLite database which includes defining main functions such as : 
  i.	insert: using the SQL insert into query values of title, author, year, and ISBN are defined and inserted into the table. 
ii.	view: executes a select statement and fetches all the rows to be viewed
iii.	search: executes a SELECT WHERE statement so the selected data can be found depending on the input that the user has given and 
         fetches the data that has been selected for. 
iv.	 delete: executes the DELETE FROM statement and the Id it has been given. 
v.	update: executes the UPDATE SET WHERE statement based on the input that the user has given. 
vi.	Def __del__(self): closes the connection after each of the functions described above has been executed. 


Credits :
Python Mega Course: Ardit Sulce